140917 database impl - not working currently
