+++ Data.py
@@ -1,25 +1,10 @@
-import json
-import logging
-
-
-import Utility
-from Clickable import Clickable
-from Config import Config
-from DataActivity import DataActivity
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
-
-
 class Data(object):
-    def __init__(self, appname, packname, _data_activity=None):
-        self.appname = appname
-        self.packname = packname
-        self.app_description = None
-        self.category = None
+    def __init__(self, _appname, _packname, _app_description=None, _category=None, _data_activity=None):
+        self.appname = _appname
+        self.packname = _packname
+        self.app_description = _app_description
+        self.category = _category
         self.data_activity = [] if _data_activity is None else _data_activity
-
-    def __str__(self):
-        return json.dumps(self, default=lambda o: o.__dict__)
 
     @staticmethod
     def encode_data(data):
@@ -29,35 +14,6 @@ class Data(object):
     @staticmethod
     def decode_data(document):
         assert document['_type'] == 'data'
-        return Data(document['appname'], document['packname'], _data_activity=document['data_activity'])
-
-    def add_new_activity(self, device, current_state=None, _click_els=None):
-        if current_state is None:
-            current_state = Utility.get_state(device)
-        # Check if state in data_activity if not, add
-        da = DataActivity(current_state)
-        # click_els = device(clickable='true', packageName=Config.pack_name) if _click_els is None else _click_els
-        click_els = device(clickable='true') if _click_els is None else _click_els
-        parent_map = Utility.create_child_to_parent(dump=device.dump())
-        for btn in click_els:
-            key = Utility.btn_to_key(btn)
-            # da.clickables.append(
-            #     Clickable(name=key, _parent_activity_state=current_state, _parent_name=Utility.xml_btn_to_key(
-            #         Utility.get_parent(btn, _parent_map=parent_map)), _siblings=Utility.get_sibling(dededededede)))
-            da.clickables.append(
-                Clickable(name=key, _parent_activity_state=current_state, _parent_name=Utility.xml_btn_to_key(
-                    Utility.get_parent(btn, _parent_map=parent_map))))
-            da.clickables_score.append(1)
-        self.data_activity.append(da)
-        logger.info('Added new activity to data.')
-
-    def to_json(self):
-        return json.dumps(self, default=lambda o: o.__dict__)
-
-    def get_activity_by_state(self, state):
-        logger.info('Getting activity by state: ' + state)
-        for activity in self.data_activity:
-            if activity.activity_state == state:
-                return activity
-        return None
-
+        return Data(_appname=document['appname'], _packname=document['packname'],
+                    _app_description=document['app_description'], _category=document['category'],
+                    _data_activity=document['data_activity'])
